requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-community
langchain_huggingface
sentence-transformers
faiss-cpu
langchain-google-genai
google-generativeai
pypdf
pypdfium2
pymupdf
unstructured[local-inference]
langchain-community[retrievers]
chromadb
lark
llama-cpp-python